Generation of the typing effect for the welcome message in the portfolio and blog
The JavaScript file is created to achieve the typing effect for the welcome message in the portfolio and blog.
Adjustments are made to the relevant style rules to achieve the mentioned effect on the specified pages.
The JavaScript file 'typewriter.js' is invoked in the corresponding HTML files to achieve the effect on the specified pages.