The content of the 'Projects' section is improved
An introduction is added for each project in the 'Projects' section.
The horizontal scroll bar is hidden.
Buttons '>' and '<' are created to enable scrolling of project cards to the right and left.
Adjustments are made in 'index.html' to add the buttons.
The buttons are styled in 'index.css'.
The 'carousel-controls.js' file is generated with the required logic to scroll left and right using the mentioned buttons.