Structural and style change in the Blog
A complete change is made in the blog's structure and styles.
The 'typewriter.js' file is adjusted because a welcome message is no longer needed on the blog.
Comments are organized, and the 'index.html' and 'index.css' files are cleaned up.
The profile picture in the 'About me' section is changed.