Project organization
The files 'index-en.html' and 'blog-en.html' are renamed to reference the 'lang' attribute for the English language.
The 'images' folder is created to store all images used in the project.
The image used in the blog's welcome section is no longer tracked, as it will not be used in this part of the project.