Structure and style generation for the publication of the first article
The cover, description, and link for the first article to be published are loaded on the blog page.
The file 'fftw-cplusplus-en.html' is generated for publishing the first blog post.
The file 'posts.css' is created to apply styles to all posts that will be published on the blog.
All links on the publication page for the first article are verified for operability.