Styles update and project structure optimization
The navigation panel styles on the Portfolio page and the first article to be published on the blog are improved by removing negative values from 'padding' and 'margin' declarations.
A folder is created for article images.
A folder is created for social media images in the 'Contact' section.
A folder is created for news images in the 'Newsletters' section.
Paths are updated in the corresponding files.