The publication of the first article begins
CSS rules in 'blog.css' and 'curriculum-vitae.css' are cleaned up.
The file 'fftw-cplusplus-en.html' is generated for article structure, along with the CSS stylesheet 'posts.css'.
Figure 1 is removed from the article and replaced with a GIF.
Adjustments are made to the 'typewriter.js' JavaScript code to create a typewriting effect on the article cover.